Stage 9 Slice 9: add Columns and RowsAffected helpers

+++ result.go
@@ -6,6 +6,6 @@ type Result struct {
 }
 
 // RowsAffected reports the number of rows changed by a write operation.
-func (r Result) RowsAffected() int64 {
-	return r.rowsAffected
+func (r Result) RowsAffected() int {
+	return int(r.rowsAffected)
 }
